+++ WorkOrderDesk.Application/Abstractions/IWorkOrderrepository.cs
@@ -1,0 +1,8 @@
+using WorkOrderDesk.Domain.WorkOrders;
+
+namespace WorkOrderDesk.Application.Abstractions;
+
+public interface IWorkOrderRepository
+{
+    Task AddAsync(WorkOrder workOrder, CancellationToken cancellationToken = default);
+}