+++ WorkOrderDesk.Application/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace WorkOrderDesk.Application;
-
-public class Class1
-{
-
-}